strange behaviour in CR
git-svn-id: svn://svn.code.sf.net/p/choco/code/trunk@1487 5779b9a9-2447-0410-887e-bb98d57c6de4

+++ choco-solver/choco-cp/src/main/java/choco/cp/model/managers/constraints/global/MultiCostRegularManager.java
@@ -62,12 +62,12 @@ int n;
                 System.arraycopy(all, 0, vs, 0, vs.length);
                 System.arraycopy(all, vs.length, z, 0, all.length - vs.length);
 
-                if (param[2] instanceof double[][][])
+                if (param[2] instanceof int[][][])
                 {
                     int[][][] csts = (int[][][]) param[2];
                     return new FastMultiCostRegular(vs,z,pi,csts, solver.getEnvironment());
                 }
-                else if (param[2] instanceof double[][][][])
+                else if (param[2] instanceof int[][][][])
                 {
                     int[][][][] csts = (int[][][][]) param[2];
                     return new FastMultiCostRegular(vs,z,pi,csts, solver.getEnvironment());
